Extract validateObjectMetadata to a validate package ✂
It doesn't depend on the API resources and can be reused and share
across multiple api version.
